Refactoring game
The three game classes were badly refactored, they still shared most of the logic and reused code.

I have refactored the base game and rounds game only.
