+++ webapp/src/components/game/BaseGameLayout.js
@@ -1,41 +1,203 @@
 "use client"
 
-import { Toolbar, Typography, Button, Card, CardContent, Grid, Box, Dialog, DialogActions, DialogContent, DialogTitle } from "@mui/material"
-import HelpOutlineIcon from "@mui/icons-material/HelpOutline"
-import PhoneIcon from "@mui/icons-material/Phone"
-import ChatIcon from "@mui/icons-material/Chat"
-import EmojiEventsIcon from "@mui/icons-material/EmojiEvents"
-import MonetizationOnIcon from "@mui/icons-material/MonetizationOn"
-import InterpreterModeIcon from '@mui/icons-material/InterpreterMode';
-import Chat from "../LLMChat"
+import { useEffect, useState } from "react"
+import { useNavigate } from "react-router";
+import { Toolbar, Typography, Button, Card, CardContent, Grid, Box, Dialog, DialogActions, DialogContent, DialogTitle, CircularProgress, Container } from "@mui/material";
+import { HelpOutline, Phone, Chat, InterpreterMode, EmojiEvents, MonetizationOn } from "@mui/icons-material";
+import LLMChat from "../LLMChat";
 import GraphComponent from '../lifelines/GraphComponent';
-import CallFriend from "../lifelines/CallFriend"
+import CallFriend from "../lifelines/CallFriend";
 import PhoneDialog from "../phone/PhoneDialog";
 import { NavLink } from "react-router";
-import { GameContainer, StyledAppBar, LogoButton, ScoreChip, CoinsChip, LifelineButton } from "./GameStyles";
+import { GameContainer, StyledAppBar, LogoButton, ScoreChip, CoinsChip, LifelineButton, LoadingContainer, ImageContainer, HighlightedTopic, OptionButton } from "./GameStyles";
+import useAxios from "../../hooks/useAxios";
+import useAuth from "../../hooks/useAuth";
+import useCoinHandler from "../../handlers/CoinHandler";
+import useLifeLinesHandler from "../../handlers/LifeLinesHandler";
+import { TOPIC_QUESTION_MAP } from "../../utils/topicQuestionMap";
 
-const BaseGame = ({
-  children,
-  roundData,
-  chatKey,
-  score,
-  correctAnswers,
-  totalRounds,
-  handleNewGame,
-  coins,
-  spentCoins,
-  showStatistics,
-  setShowStatistics,
-  canAfford,
-  handleFiftyFifty,
-  handleCallFriend,
-  handleCloseCallFriend,
-  handleAudienceCall,
-  handlePhoneOut,
-  handlePhoneOutClose,
-  handleUseChat,
-  isTrue
-}) => {
+const BaseGame = ({ children, onNewGame, onRoundComplete, gameEnding }) => {
+  const axios = useAxios();
+  const { auth } = useAuth();
+  const navigate = useNavigate();
+
+  const [roundData, setRoundData] = useState(null);
+  const [score, setScore] = useState(0);
+  const [selectedAnswer, setSelectedAnswer] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [chatKey, setChatKey] = useState(0); // resets the chat component every time it is updated
+  const [showStatistics, setShowStatistics] = useState(false);
+  const [questions, setQuestions] = useState([]);
+  const [correctAnswers, setCorrectAnswers] = useState(0);
+  const [roundsPlayed, setRoundsPlayed] = useState(0);
+
+  const { coins, spentCoins, setSpentCoins, canAfford, spendCoins, fetchUserCoins, updateUserCoins } = useCoinHandler(axios, auth);
+  const { handleFiftyFifty, handleCallFriend, handleCloseCallFriend, handleAudienceCall, handlePhoneOut, handlePhoneOutClose, handleUseChat,
+    hiddenOptions, isTrue, setHiddenOptions, setShowGraph, newGame } = useLifeLinesHandler(roundData, spendCoins);
+
+  // Set up the game when the component mounts
+  useEffect(() => {
+    gameSetup();
+  }, []);
+
+  useEffect(() => {
+    fetchUserCoins();
+  }, [auth]);
+
+  const gameSetup = async () => {
+    try {
+      // First round
+      const data = await loadRound()
+      setRoundData(data)
+    } catch (error) {
+      console.error("Error fetching data from question service:", error)
+      setLoading(false)
+    }
+  }
+
+  // Function to load the data for each round.
+  const loadRound = async () => {
+    try {
+      setLoading(true)
+      setChatKey(chatKey + 1);
+      setHiddenOptions([])
+
+      const selectedTopics = JSON.parse(sessionStorage.getItem('selectedTopics'));
+      let response = "";
+
+      if (!selectedTopics || !Array.isArray(selectedTopics) || selectedTopics.length === 0) {
+        navigate("/home", { replace: true });
+      } else {
+        response = await axios.get("/getRound", {
+          params: { topics: selectedTopics }
+        });
+      }
+
+      return response.data
+    } catch (error) {
+      console.error("Error fetching data from question service:", error)
+      setLoading(false)
+    }
+  }
+
+  // Check if the game is still loading after modifying the round data
+  useEffect(() => {
+    if (roundData && roundData.items.length > 0) {
+      const questionInfo = TOPIC_QUESTION_MAP[roundData.topic] || { wh: "What", name: roundData.topic }; // Value by default if topic is not found
+      //setRoundPrompt(`${questionInfo.wh} is this ${questionInfo.name}?`);
+      setLoading(false);
+    } else {
+      setLoading(true);
+    }
+  }, [roundData]);
+
+  // Load data every 500ms while the game is loading
+  useEffect(() => {
+    let intervalId;
+
+    if (loading) {
+      intervalId = setInterval(async () => {
+        try {
+          const data = await loadRound();
+          if (data && data.items.length > 0) {
+            setRoundData(data);
+          }
+        } catch (error) {
+          console.error("Error in interval loading round data:", error);
+        }
+      }, 500);
+    }
+
+    return () => {
+      if (intervalId) {
+        clearInterval(intervalId); // Cleanup interval on unmount or when loading stops
+      }
+    };
+  }, [loading]);
+
+  // Add the game statistics to the database and show the statistics dialog
+  const endGame = async () => {
+    try {
+      // Calculate earned coins based on score
+      const earnedCoins = Math.round(score * 0.3);
+      await updateUserCoins(earnedCoins);
+      await axios.post("/addgame", { username: auth.username, questions });
+    } catch (error) {
+      console.error("Error saving user stadistics:", error);
+    }
+
+    setShowStatistics(true);
+  };
+
+  const handleNewGame = async () => {
+    setShowStatistics(false);
+    setRoundData(null);
+    setScore(0);
+    newGame();
+    setSelectedAnswer(null);
+    fetchUserCoins(); // Reset coins to the latest value from the server
+    setSpentCoins(0);
+    setQuestions([]);
+    setCorrectAnswers(0);
+
+    onNewGame();
+
+    gameSetup();
+  };
+
+  const handleOptionSelect = async (index) => {
+    if (selectedAnswer !== null) return;
+
+    const isCorrect = correctOption(index);
+    setSelectedAnswer(index);
+
+    const pointsIncrement = 50;
+    if (isCorrect) {
+      setScore(score + pointsIncrement);
+      setCorrectAnswers(correctAnswers + 1);
+      setShowGraph(false);
+    }
+
+    let updatedQuestions = [];
+    setQuestions((prev) => {
+      updatedQuestions = [
+        ...prev,
+        {
+          topic: roundData.topic,
+          isCorrect: isCorrect,
+          pointsIncrement: pointsIncrement,
+        },
+      ];
+      return updatedQuestions;
+    });
+
+    setTimeout(async () => {
+      setRoundsPlayed(roundsPlayed + 1);
+
+      if (gameEnding()) {
+        return await endGame(updatedQuestions);
+      }
+
+      onRoundComplete();
+      setRoundData(null);
+      setSelectedAnswer(null);
+
+      try {
+        const data = await loadRound();
+        setRoundData(data);
+      } catch (error) {
+        console.error("Error loading new round", error);
+        setLoading(false);
+      }
+    }, 2000);
+  }
+
+  const correctOption = (index) => {
+    if (!roundData) return false
+    const selectedName = roundData.items[index].name
+    const correctName = roundData.itemWithImage.name
+    return selectedName === correctName
+  }
 
   return (
     <GameContainer maxWidth="100%" height="100%">
@@ -47,11 +209,11 @@ const BaseGame = ({
           </LogoButton>
           <Box sx={{ flexGrow: 1 }} />
           <Box sx={{ display: "flex", gap: 2 }}>
-            <CoinsChip variant="contained" startIcon={<MonetizationOnIcon />}>
+            <CoinsChip variant="contained" startIcon={<MonetizationOn />}>
               {coins} 🪙
             </CoinsChip>
             <ScoreChip elevation={0}>
-              <EmojiEventsIcon sx={{ mr: 1 }} />
+              <EmojiEvents sx={{ mr: 1 }} />
               Score: {score}
             </ScoreChip>
           </Box>
@@ -69,7 +231,7 @@ const BaseGame = ({
               </Typography>
               <LifelineButton
                 variant="contained"
-                startIcon={<HelpOutlineIcon />}
+                startIcon={<HelpOutline />}
                 onClick={handleFiftyFifty}
                 disabled={isTrue("50") || !canAfford(100)}
                 isUsed={isTrue("50")}
@@ -80,7 +242,7 @@ const BaseGame = ({
 
               <LifelineButton
                 variant="contained"
-                startIcon={<PhoneIcon />}
+                startIcon={<Phone />}
                 onClick={handleCallFriend}
                 disabled={isTrue("CallFriend")}
                 isUsed={isTrue("CallFriend")}
@@ -91,7 +253,7 @@ const BaseGame = ({
 
               <LifelineButton
                 variant="contained"
-                startIcon={<InterpreterModeIcon />}
+                startIcon={<InterpreterMode />}
                 onClick={handleAudienceCall}
                 disabled={isTrue("AskAudience") || !canAfford(150)}
                 isUsed={isTrue("AskAudience")}
@@ -102,7 +264,7 @@ const BaseGame = ({
 
               <LifelineButton
                 variant="contained"
-                startIcon={<ChatIcon />}
+                startIcon={<Chat />}
                 onClick={handleUseChat}
                 disabled={isTrue("UseChat") || !canAfford(200)}
                 isUsed={isTrue("UseChat")}
@@ -132,54 +294,107 @@ const BaseGame = ({
               />)}
 
             </CardContent>{isTrue("ShowGraph") && (
-            <Card elevation={3} sx={{ marginTop: 2, paddingTop: 3 }}>
-              <CardContent>
-                <Typography variant="h4" component="h2" color="primary" sx={{ fontSize: '1.5rem' }}>
-                  The audience says...
-                </Typography>
-                {roundData && <GraphComponent correctAnswer={roundData.itemWithImage.name}
-                  distractors={roundData.items
-                    .filter(item => item.name !== roundData.itemWithImage.name)
-                    .map(item => item.name)
-                  }
-                />}
-              </CardContent>
-            </Card>
-          )}
+              <Card elevation={3} sx={{ marginTop: 2, paddingTop: 3 }}>
+                <CardContent>
+                  <Typography variant="h4" component="h2" color="primary" sx={{ fontSize: '1.5rem' }}>
+                    The audience says...
+                  </Typography>
+                  {roundData && <GraphComponent correctAnswer={roundData.itemWithImage.name}
+                    distractors={roundData.items
+                      .filter(item => item.name !== roundData.itemWithImage.name)
+                      .map(item => item.name)
+                    }
+                  />}
+                </CardContent>
+              </Card>
+            )}
           </Card>
-          
+
         </Grid>
 
         {/* Game Area */}
-        {children}
+        <Grid item xs={12} md={6}>
+          <Card elevation={3} sx={{ height: "100%", minHeight: 400 }}>
+            <CardContent>
+              {loading ? (
+                <LoadingContainer>
+                  <CircularProgress size={60} thickness={4} color="primary" />
+                  <Typography variant="h6" color="textSecondary" sx={{ mt: 3 }}>
+                    Loading question...
+                  </Typography>
+                  <Typography variant="body2" color="textSecondary" sx={{ mt: 1 }}>
+                    Get ready for round {roundsPlayed + 1}!
+                  </Typography>
+                </LoadingContainer>
+              ) : (
+                roundData && (
+                  <>
+                    {children}
+                    <ImageContainer>
+                      <img
+                        src={roundData.itemWithImage.imageUrl || "/placeholder.svg"}
+                        alt={roundData.itemWithImage.imageAltText || "Item image"}
+                      />
+                    </ImageContainer>
+                    <Container sx={{ textAlign: "center", mb: 2 }}>
+                      {roundData && (
+                        <Typography data-testid="question-prompt" variant="h5" gutterBottom sx={{ fontWeight: "bold" }}>
+                          {TOPIC_QUESTION_MAP[roundData.topic]?.wh || "What"} is this <HighlightedTopic>{TOPIC_QUESTION_MAP[roundData.topic]?.name || roundData.topic.toUpperCase()}</HighlightedTopic>?
+                        </Typography>
+                      )}
+                    </Container>
+                    <Grid container spacing={2}>
+                      {roundData.items.map((item, index) => (
+                        <Grid item xs={6} key={index}>
+                          <OptionButton
+                            variant="contained"
+                            fullWidth
+                            onClick={() => handleOptionSelect(index)}
+                            disabled={hiddenOptions.includes(index)}
+                            isHidden={hiddenOptions.includes(index)}
+                            hasSelectedAnswer={selectedAnswer !== null}
+                            isSelected={selectedAnswer === index}
+                            isCorrect={correctOption(index)}
+                          >
+                            {item.name}
+                          </OptionButton>
+                        </Grid>
+                      ))}
+                    </Grid>
+                  </>
+                )
+              )}
+            </CardContent>
+          </Card>
+        </Grid>
 
         {/* Right Side (Chat) */}
         <Grid item xs={12} md={3}>
           <Card elevation={3} sx={{ height: "100%" }}>
             <CardContent>
-              {roundData && <Chat key={chatKey} roundData={roundData} />}
+              {roundData && <LLMChat key={chatKey} roundData={roundData} />}
             </CardContent>
           </Card>
         </Grid>
       </Grid>
 
       {/* Game statistics */}
-      <Dialog 
-        open={showStatistics} 
+      <Dialog
+        open={showStatistics}
         onClose={(event, reason) => {
           // Prevent the user to interact with the rest of the screen when the dialog is shown
           if (reason !== "backdropClick" && reason !== "escapeKeyDown") {
             setShowStatistics(false)
           }
         }}
-        >
+      >
         <DialogTitle sx={{ fontWeight: "bold", color: "primary.main" }}>Game Over</DialogTitle>
         <DialogContent>
           <Typography variant="body1"><b>Final Score:</b> {score}</Typography>
-          <Typography variant="body1"><b>Correct Answers:</b> {correctAnswers} / {totalRounds}</Typography>
-          <Typography variant="body1"><b>Accuracy Rate:</b> {((correctAnswers / totalRounds) * 100).toFixed(2)}%</Typography>
+          <Typography variant="body1"><b>Correct Answers:</b> {correctAnswers} / {roundsPlayed}</Typography>
+          <Typography variant="body1"><b>Accuracy Rate:</b> {((correctAnswers / roundsPlayed) * 100).toFixed(2)}%</Typography>
           <Typography variant="body1"><b>Spent on lifelines:</b> {spentCoins} 🪙</Typography>
-          <Typography variant="body1"><b>Coins earned:</b> {Math.round(score * 0.3)} 🪙</Typography>     
+          <Typography variant="body1"><b>Coins earned:</b> {Math.round(score * 0.3)} 🪙</Typography>
         </DialogContent>
         <DialogActions sx={{ justifyContent: "center", pb: 2 }}>
           <NavLink to="/home">
